requirements.txt
arabic-reshaper==3.0.0
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.10.1
django-db-logger==0.1.13
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
itypes==1.2.0
Jinja2==3.0.3
lxml==4.9.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
openpyxl==3.0.10
Pillow==8.4.0
psycopg-binary==3.0.18
psycopg2==2.9.5
PyPDF2==3.0.1
python-bidi==0.4.2
python-docx==0.8.11
pytz==2022.7
reportlab==3.6.6
requests==2.27.1
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
xhtml2pdf==0.2.5
XlsxWriter==3.1.0
xmltodict==0.13.0
